scripts/requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
requests==2.32.2
scikit-learn==1.4.2
google-api-python-client==2.130.0
google-auth-oauthlib==1.2.0
pyautogui==0.9.54
pyvirtualdisplay==3.0
chromium-chromedriver
